Split `studies` app models into separate files
This resolves a bunch of circular import issues, which required function-local
imports.
